+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Release notes for Agda version 2.6.1
 ====================================
 
+
+Pragmas and options
+-------------------
+
+* New pragma `WARNING_ON_IMPORT` to let module authors raise a warning
+  when a module is imported. This can be use to tell users deprecations.
+
 Release notes for Agda version 2.6.0
 ====================================